+++ src/commands/leaderboard.ts
@@ -1,6 +1,6 @@
 import Discord, { MessageActionRow, MessageButton } from 'discord.js'
 import Command from '../struct/Command.js'
-import User from '../struct/User.js'
+import User from '../struct/Builder.js'
 
 export default new Command({
     name: 'leaderboard',
